proyecto/src/backend/requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-cors-headers>=4.4
python-dotenv>=1.0

# Database
psycopg2-binary>=2.9

# Task Queue
celery>=5.4
redis>=5.0
django-redis>=5.4
django-celery-beat>=2.5
django-celery-results>=2.5

# Storage
django-storages>=1.14
boto3>=1.34

# Testing
pytest>=8.0
pytest-django>=4.8

# System Monitoring
psutil>=5.9

# Production Server
gunicorn>=21.2

# PDF Processing
PyPDF2>=3.0
reportlab>=4.0

# Web Push Notifications
pywebpush>=1.14

# Timezone
pytz>=2024.1

# Web Scraping
beautifulsoup4>=4.12
requests>=2.31

# Google OAuth
google-auth>=2.40
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2
google-api-python-client>=2.180

# Markdown Support
markdown>=3.4

# Optional (commented out)
# django-channels>=4.1
# channels-redis>=4.2
